requirements.txt
torch>=1.12.0
torchvision>=0.13.0
numpy>=1.21.0,<2.0.0
opencv-python>=4.5.0
pillow>=9.0.0
pyyaml>=6.0
tensorboard>=2.9.0
tqdm>=4.64.0
onnx>=1.12.0
scikit-image>=0.19.0

# Optional: for MAC counting
thop

# Optional: for TFLite export
# onnx-tf
# tensorflow>=2.10.0
